+++ code/include/dummyframework/network/udpserver.h
@@ -61,11 +61,11 @@ namespace DummyFramework
 
 		bool attach(udpconnection& conn);
 		bool accept(udpchannel& out);
-		bool addpacket(unsigned long client_id, const udppacket& packet);
-		bool getpacket(unsigned long client_id, udppacket& packet);
+		bool addpacket(unsigned int client_id, const udppacket& packet);
+		bool getpacket(unsigned int client_id, udppacket& packet);
 		void run();
 
-		inline void getaddress(unsigned long client_id, ipv4address& addr) {
+		inline void getaddress(unsigned int client_id, ipv4address& addr) {
 			addr = clients[client_id].address;
 		}
 
@@ -77,5 +77,3 @@ namespace DummyFramework
 
 #endif
 //=============================================================================================================
-
-
